Update README.md
Blood bank and donor mangement system
Web-based Blood Bank and Donor Management System is a management system developed as a web-based portal that enables individuals who want to donate blood to help the needy.

It also enables hospitals to record and store the data for people who want to communicate with them, and it also provides a centralized blood bank database. The system is developed by using HTML, PHP, and MySQL as a database system to manage and store the data. The Waterfall Methodology, which is the traditional version and the classic approach of a system development life cycle, is used to develop and build the webbased blood bank.

The system targets three types of user: the public who wants to donate blood, the recipients who need the donated blood, and the hospitals who that work as an intermediary to manage the communication between the donors and recipient.